Put Pwm struct in separate module.
